#18 refactor dir names
