+++ spring_API/src/main/java/com/BlockHeads/config/SecurityConfig.java
@@ -1,4 +1,4 @@
-package com.BlockHeads.configs;
+package com.BlockHeads.config;
 
 import org.springframework.boot.autoconfigure.security.servlet.PathRequest;
 import org.springframework.context.annotation.Bean;
