+++ spring_API/src/main/java/com/BlockHeads/repository/UserRepository.java
@@ -1,10 +1,10 @@
-package com.BlockHeads.repositories;
+package com.BlockHeads.repository;
 
 import java.util.Optional;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.BlockHeads.models.User;
+import com.BlockHeads.model.User;
 
 public interface UserRepository extends JpaRepository<User, Long> {
     Optional<User> findByEmail(String email);
